ci(test): add workflows for utils and CLI testing
