+++ .github/workflows/quality.yml
@@ -51,22 +51,33 @@ jobs:
       - name: Build packages
         run: pnpm build --affected
 
-  unit_test:
-    name: Unit Test
+  test_cli:
+    name: Test / CLI
     if: ${{ github.event_name == 'merge_group' || !github.event.pull_request.draft }}
     needs: scm_base
-    uses: ./.github/workflows/unit-test.yml
+    uses: ./.github/workflows/test-cli.yml
     secrets:
       TURBO_TOKEN: ${{ secrets.TURBO_TOKEN }}
       CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
     with:
       TURBO_SCM_BASE: ${{ needs.scm_base.outputs.sha }}
 
-  browser_test:
-    name: Browser Test
+  test_react:
+    name: Test / React
     if: ${{ github.event_name == 'merge_group' || !github.event.pull_request.draft }}
     needs: scm_base
-    uses: ./.github/workflows/browser-test.yml
+    uses: ./.github/workflows/test-react.yml
+    secrets:
+      TURBO_TOKEN: ${{ secrets.TURBO_TOKEN }}
+      CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
+    with:
+      TURBO_SCM_BASE: ${{ needs.scm_base.outputs.sha }}
+
+  test_utils:
+    name: Test / Utils
+    if: ${{ github.event_name == 'merge_group' || !github.event.pull_request.draft }}
+    needs: scm_base
+    uses: ./.github/workflows/test-utils.yml
     secrets:
       TURBO_TOKEN: ${{ secrets.TURBO_TOKEN }}
       CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
