+++ .github/workflows/test-react.yml
@@ -1,4 +1,4 @@
-name: Browser Test
+name: Test / React
 
 on:
   workflow_call:
@@ -19,24 +19,34 @@ env:
 
 jobs:
   test:
-    name: React / ${{ matrix.name }} ${{ matrix.shard_index }}/4
+    name: ${{ matrix.name }} ${{ matrix.shard_index }}/4
     runs-on: ubuntu-24.04
     timeout-minutes: 10
     strategy:
       fail-fast: false
       matrix:
-        browser: [chromium, firefox, webkit]
+        env: [jsdom, chromium, firefox, webkit]
         shard_index: [1, 2, 3, 4]
         include:
-          - browser: chromium
+          - env: jsdom
+            name: JSDOM
+            diff: |
+              - '!packages/react/src/**/*.test.{browser,chromium,firefox,webkit}.{ts,tsx}'
+          - env: chromium
             name: Chromium
-            diff: "!packages/react/src/**/*.{firefox,webkit}.{ts,tsx,mts}"
-          - browser: firefox
+            diff: |
+              - '!packages/react/src/**/*.test.{ts,tsx,mts}'
+              - '!packages/react/src/**/*.test.{firefox,webkit}.{ts,tsx,mts}'
+          - env: firefox
             name: Firefox
-            diff: "!packages/react/src/**/*.{chromium,webkit}.{ts,tsx,mts}"
-          - browser: webkit
+            diff: |
+              - '!packages/react/src/**/*.test.{ts,tsx,mts}'
+              - '!packages/react/src/**/*.test.{chromium,webkit}.{ts,tsx,mts}'
+          - env: webkit
             name: Webkit
-            diff: "!packages/react/src/**/*.{chromium,firefox}.{ts,tsx,mts}"
+            diff: |
+              - '!packages/react/src/**/*.test.{ts,tsx,mts}'
+              - '!packages/react/src/**/*.test.{chromium,firefox}.{ts,tsx,mts}'
     steps:
       - name: Checkout
         uses: actions/checkout@8e8c483db84b4bee98b60c0593521ed34d9990e8 # v6.0.1
@@ -57,32 +67,32 @@ jobs:
               - packages/react/test/**
             diff:
               - packages/react/src/**/*.{ts,tsx,mts}
-              - '!packages/react/src/**/*.{test,stories}.{ts,tsx}'
-              - '${{ matrix.diff }}'
+              - '!packages/react/src/**/*.stories.{ts,tsx}'
+              ${{ matrix.diff }}
 
       - name: Install
         if: inputs.TURBO_SCM_BASE == '' || steps.changed-files.outputs.changed_keys != ''
         uses: ./.github/composite-actions/install
 
       - name: Setup Browser
-        if: inputs.TURBO_SCM_BASE == '' || steps.changed-files.outputs.changed_keys != ''
-        run: pnpm react exec playwright install ${{ matrix.browser }} --with-deps ${{ matrix.browser == 'chromium' && '--only-shell' || '' }}
+        if: matrix.env != 'jsdom' && (inputs.TURBO_SCM_BASE == '' || steps.changed-files.outputs.changed_keys != '')
+        run: pnpm react exec playwright install ${{ matrix.env }} --with-deps ${{ matrix.env == 'chromium' && '--only-shell' || '' }}
 
       - name: Run test (full)
         if: inputs.TURBO_SCM_BASE == '' || steps.changed-files.outputs.full_any_changed == 'true'
-        run: pnpm turbo run ${{ matrix.browser == 'firefox' && 'test:firefox' || format('test:coverage:{0}', matrix.browser) }} --filter=./packages/react -- ${{ matrix.browser == 'firefox' && '--run' || '' }} --retry=3 --shard=${{ matrix.shard_index }}/4
+        run: pnpm turbo run test:coverage:${{ matrix.env }} --filter=./packages/react -- --retry=3 --shard=${{ matrix.shard_index }}/4
 
       - name: Run test (diff)
         if: inputs.TURBO_SCM_BASE != '' && steps.changed-files.outputs.full_any_changed != 'true' && steps.changed-files.outputs.diff_any_changed == 'true'
-        run: pnpm turbo run ${{ matrix.browser == 'firefox' && 'test:firefox' || format('test:coverage:{0}', matrix.browser) }} --filter=./packages/react -- ${{ matrix.browser == 'firefox' && '--run' || '' }} --retry=3 --changed ${{ inputs.TURBO_SCM_BASE }} --shard=${{ matrix.shard_index }}/4
+        run: pnpm turbo run test:coverage:${{ matrix.env }} --filter=./packages/react -- --retry=3 --changed ${{ inputs.TURBO_SCM_BASE }} --shard=${{ matrix.shard_index }}/4
 
       - name: Upload coverage
-        if: matrix.browser != 'firefox' && (inputs.TURBO_SCM_BASE == '' || steps.changed-files.outputs.changed_keys != '')
+        if: inputs.TURBO_SCM_BASE == '' || steps.changed-files.outputs.changed_keys != ''
         uses: codecov/codecov-action@75cd11691c0faa626561e295848008c8a7dddffe # v5.5.4
         with:
           disable_search: true
           fail_ci_if_error: false
-          files: ./packages/react/coverage/${{ matrix.browser }}/coverage-final.json
+          files: ./packages/react/coverage/${{ matrix.env }}/coverage-final.json
           flags: unittests
           token: ${{ secrets.CODECOV_TOKEN }}
           verbose: true
