+++ .github/workflows/test-utils.yml
@@ -1,0 +1,97 @@
+name: Test / Utils
+
+on:
+  workflow_call:
+    inputs:
+      TURBO_SCM_BASE:
+        required: false
+        type: string
+    secrets:
+      TURBO_TOKEN:
+        required: false
+      CODECOV_TOKEN:
+        required: false
+
+env:
+  TURBO_TOKEN: ${{ secrets.TURBO_TOKEN }}
+  TURBO_TEAM: ${{ vars.TURBO_TEAM }}
+  TURBO_SCM_BASE: ${{ inputs.TURBO_SCM_BASE }}
+
+jobs:
+  test:
+    name: ${{ matrix.name }}
+    runs-on: ubuntu-24.04
+    timeout-minutes: 10
+    strategy:
+      fail-fast: false
+      matrix:
+        env: [jsdom, chromium, firefox, webkit]
+        include:
+          - env: jsdom
+            name: JSDOM
+            diff: |
+              - '!packages/utils/src/**/*.test.{browser,chromium,firefox,webkit}.{ts,tsx}'
+          - env: chromium
+            name: Chromium
+            diff: |
+              - '!packages/utils/src/**/*.test.{ts,tsx,mts}'
+              - '!packages/utils/src/**/*.test.{firefox,webkit}.{ts,tsx,mts}'
+          - env: firefox
+            name: Firefox
+            diff: |
+              - '!packages/utils/src/**/*.test.{ts,tsx,mts}'
+              - '!packages/utils/src/**/*.test.{chromium,webkit}.{ts,tsx,mts}'
+          - env: webkit
+            name: Webkit
+            diff: |
+              - '!packages/utils/src/**/*.test.{ts,tsx,mts}'
+              - '!packages/utils/src/**/*.test.{chromium,firefox}.{ts,tsx,mts}'
+    steps:
+      - name: Checkout
+        uses: actions/checkout@8e8c483db84b4bee98b60c0593521ed34d9990e8 # v6.0.1
+        with:
+          fetch-depth: 0
+
+      - name: Check modified files
+        id: changed-files
+        uses: tj-actions/changed-files@9426d40962ed5378910ee2e21d5f8c6fcbf2dd96 # v47.0.6
+        with:
+          files_yaml: |
+            full:
+              - .github/workflows/quality.yml
+              - '{turbo.json,package.json}'
+              - packages/workspace/package.json
+              - packages/workspace/src/vitest/**
+              - packages/utils/{package.json,vitest.config.ts}
+            diff:
+              - packages/utils/src/**/*.{ts,tsx,mts}
+              - '!packages/utils/src/**/*.stories.{ts,tsx}'
+              ${{ matrix.diff }}
+
+      - name: Install
+        if: inputs.TURBO_SCM_BASE == '' || steps.changed-files.outputs.changed_keys != ''
+        uses: ./.github/composite-actions/install
+
+      - name: Setup Browser
+        if: matrix.env != 'jsdom' && (inputs.TURBO_SCM_BASE == '' || steps.changed-files.outputs.changed_keys != '')
+        run: pnpm utils exec playwright install ${{ matrix.env }} --with-deps ${{ matrix.env == 'chromium' && '--only-shell' || '' }}
+
+      - name: Run test (full)
+        if: inputs.TURBO_SCM_BASE == '' || steps.changed-files.outputs.full_any_changed == 'true'
+        run: pnpm turbo run test:coverage:${{ matrix.env }} --filter=./packages/utils -- --retry=3
+
+      - name: Run test (diff)
+        if: inputs.TURBO_SCM_BASE != '' && steps.changed-files.outputs.full_any_changed != 'true' && steps.changed-files.outputs.diff_any_changed == 'true'
+        run: pnpm turbo run test:coverage:${{ matrix.env }} --filter=./packages/utils -- --retry=3 --changed ${{ inputs.TURBO_SCM_BASE }}
+
+      - name: Upload coverage
+        if: inputs.TURBO_SCM_BASE == '' || steps.changed-files.outputs.changed_keys != ''
+        uses: codecov/codecov-action@75cd11691c0faa626561e295848008c8a7dddffe # v5.5.4
+        with:
+          disable_search: true
+          fail_ci_if_error: false
+          files: ./packages/utils/coverage/${{ matrix.env }}/coverage-final.json
+          flags: unittests
+          token: ${{ secrets.CODECOV_TOKEN }}
+          verbose: true
+          codecov_yml_path: ./codecov.yml
