+++ packages/react/package.json
@@ -135,12 +135,12 @@
     "@types/jsdom": "^28.0.1",
     "@types/yamljs": "catalog:",
     "@vitejs/plugin-react": "catalog:",
-    "@vitest/browser": "^4.1.5",
-    "@vitest/browser-playwright": "^4.1.5",
+    "@vitest/browser": "catalog:",
+    "@vitest/browser-playwright": "catalog:",
     "@yamada-ui/workspace": "workspace:*",
     "axe-playwright": "^2.2.2",
     "jsdom": "^29.1.1",
-    "playwright": "^1.59.1",
+    "playwright": "catalog:",
     "react-hook-form": "catalog:",
     "react-icons": "catalog:",
     "use-mask-input": "^3.9.0",
