+++ packages/utils/package.json
@@ -33,8 +33,22 @@
     "clean": "rimraf node_modules dist trace coverage .turbo",
     "build": "tsdown",
     "test": "vitest",
-    "test:coverage": "vitest run --coverage",
+    "test:jsdom": "vitest --project=jsdom",
+    "test:browser": "vitest --project=browser:chromium --project=browser:firefox --project=browser:webkit",
+    "test:chromium": "vitest --project=browser:chromium",
+    "test:webkit": "vitest --project=browser:webkit",
+    "test:firefox": "vitest --project=browser:firefox",
     "test:dev": "vitest --watch --ui",
+    "test:dev:jsdom": "vitest --watch --ui --project=jsdom",
+    "test:dev:browser": "vitest --watch --ui --project=browser:chromium --project=browser:firefox --project=browser:webkit",
+    "test:dev:chromium": "vitest --watch --ui --project=browser:chromium",
+    "test:dev:webkit": "vitest --watch --ui --project=browser:webkit",
+    "test:dev:firefox": "vitest --watch --ui --project=browser:firefox",
+    "test:coverage": "vitest run --coverage",
+    "test:coverage:jsdom": "vitest run --project=jsdom --coverage --coverage.reportsDirectory=coverage/jsdom",
+    "test:coverage:chromium": "vitest run --project=browser:chromium --coverage --coverage.provider=istanbul --coverage.reportsDirectory=coverage/chromium",
+    "test:coverage:webkit": "vitest run --project=browser:webkit --coverage --coverage.provider=istanbul --coverage.reportsDirectory=coverage/webkit",
+    "test:coverage:firefox": "vitest run --project=browser:firefox --coverage --coverage.provider=istanbul --coverage.reportsDirectory=coverage/firefox",
     "typecheck": "tsgo",
     "typetrace": "tsgo --generateTrace ./trace --incremental false",
     "typeanalyze": "npx analyze-trace ./trace --expandTypes=false",
@@ -48,7 +62,10 @@
     "color2k": "^2.0.3"
   },
   "devDependencies": {
+    "@vitest/browser": "catalog:",
+    "@vitest/browser-playwright": "catalog:",
     "@yamada-ui/workspace": "workspace:*",
+    "playwright": "catalog:",
     "react": "catalog:"
   },
   "peerDependencies": {
