+++ packages/utils/src/dom.test.chromium.ts
@@ -1,0 +1,9 @@
+import { isChrome } from "./dom"
+
+describe("DOM", () => {
+  describe("isChrome", () => {
+    test("should return true in Chromium", () => {
+      expect(isChrome()).toBe(true)
+    })
+  })
+})
