+++ packages/utils/src/dom.test.firefox.ts
@@ -1,0 +1,9 @@
+import { isFirefox } from "./dom"
+
+describe("DOM", () => {
+  describe("isFirefox", () => {
+    test("should return true in Firefox", () => {
+      expect(isFirefox()).toBe(true)
+    })
+  })
+})
