+++ packages/utils/src/dom.test.webkit.ts
@@ -1,0 +1,15 @@
+import { isSafari, isWebKit } from "./dom"
+
+describe("DOM", () => {
+  describe("isWebKit", () => {
+    test("should return true in WebKit", () => {
+      expect(isWebKit()).toBe(true)
+    })
+  })
+
+  describe("isSafari", () => {
+    test("should return true in WebKit", () => {
+      expect(isSafari()).toBe(true)
+    })
+  })
+})
